requirements.txt
moderngl
numpy
pygame
PyGLM
pywavefront
opencv-python
opencv-contrib-python
os
PyYAML
pytesseract
scipy
